Separate Logic in Different files
